% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curvefit.R
\name{gr_cf_fit_add}
\alias{gr_cf_fit_add}
\title{Add peaks to a spectrum data frame}
\usage{
gr_cf_fit_add(spectrum_file, cf_file)
}
\arguments{
\item{spectrum_file}{Path to a tab-separated file with a spectrum in two columns (wavenumber and intensity)}

\item{cf_file}{Path to a file containing a curvefit results table (made in Wire by right-clicking results of a curvefit and choosing Copy Results)}
}
\description{
Supplied with a spectrum as a tab-separated text-file with wavenumber and intensites, and a curvefit-results table from Wire, it will generate columns for each peak with the corresponding intensities, for easy plotting. The envelope of the fit is also calculated by summing all peaks at each wavenumber.
}
\examples{
curvefit <- system.file("extdata/curvefit_data.txt", package = "gRaphene")
spectrum <- system.file("extdata/spectrum_example.txt", package = "gRaphene")

gr_cf_fit_add(spectrum, curvefit)

}
\seealso{
Other raman fits: \code{\link{gr_cf_fit_plot}}
}
