% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curvefit.R
\name{gr_cf_fit_plot}
\alias{gr_cf_fit_plot}
\title{Plot fitted spectrum with peaks}
\usage{
gr_cf_fit_plot(spec, curvefit)
}
\description{
Takes a spectrum and corresponding curvefit results and generates a plot
}
\examples{

curvefit <- system.file("extdata/curvefit_data.txt", package = "gRaphene")
spectrum <- system.file("extdata/spectrum_example.txt", package = "gRaphene")

gr_cf_fit_plot(spectrum, curvefit)

}
\seealso{
Other raman fits: \code{\link{gr_cf_fit_add}}
}
