% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{gr_cf_nest}
\alias{gr_cf_nest}
\title{Return a nested tibble contaning all curve fits}
\usage{
gr_cf_nest(df)
}
\arguments{
\item{df}{A dataframe from \code{gr_cf_read}}
}
\description{
This is mostly an internal function. The nested tibble is used by \code{gr_cf_plot_all()}.
}
