% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{gr_cf_plot_all}
\alias{gr_cf_plot_all}
\title{Plot all curve fit maps at once}
\usage{
gr_cf_plot_all(df, scalebar = 1, palette = "Spectral", ncol = 3)
}
\arguments{
\item{df}{A curve fit data frame, as generated by \code{gr_cf_read}}

\item{scalebar}{Size of the scalebar (defaults to 1 um)}

\item{palette}{Color palette used in the map. Defaults to Spectral. See `RColorBrewer::display.brewer.all()` for more options}

\item{ncol}{Number of columns in the generated grid.}
}
\description{
All curve fit maps in \code{df} are plotted in a grid.
}
\seealso{
Other curve fit functions: \code{\link{gr_cf_plot}},
  \code{\link{gr_cf_read}},
  \code{\link{gr_cf_summary_save}},
  \code{\link{gr_cf_summary}}
}
