% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{gr_cf_summary}
\alias{gr_cf_summary}
\title{Create summary of a graphene curve fit data frame}
\usage{
gr_cf_summary(df)
}
\arguments{
\item{df}{Computes summary (min, mean, max and sd) for \code{df}}
}
\description{
Takes the results of gRdata_read and computes summary statistics
}
\examples{
gr_cf_summary(df_cf)
}
\seealso{
Other curve fit functions: \code{\link{gr_cf_plot_all}},
  \code{\link{gr_cf_plot}}, \code{\link{gr_cf_read}},
  \code{\link{gr_cf_summary_save}}
}
\keyword{summary}
