% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curvefit.R
\name{gr_gaussian}
\alias{gr_gaussian}
\title{Gaussian curve}
\usage{
gr_gaussian(x, G, x_0, A)
}
\arguments{
\item{x}{Point along x-axis, e.g. the wavenumber}

\item{G}{Full-width at half max of the peak}

\item{x_0}{Center of the peak}

\item{A}{Area of the peak}
}
\description{
Generates a Gaussian curve based on the input data.
}
\examples{
x <- 1:1000
plot(x, gr_gaussian(x, G = 10, x_0 = 500, A = 10))

}
\references{
\url{http://mathworld.wolfram.com/GaussianFunction.html}
}
