% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{gr_map_read}
\alias{gr_map_read}
\title{Load complete spectrum map from exported txt file}
\usage{
gr_map_read(path)
}
\arguments{
\item{path}{Path to txt file containing the spectrum map}
}
\description{
Load all spectra and corresponding (x,y) coordinates and wrangles into a tidy tibble
Expects a tab-separated file with 4 columns (5 if z-value is present - not implemented yet!)
}
\examples{
gr_map_read("graphene_map.txt")
gr_map_read(system.file('extdata', 'graphene_spectrum_map_export.txt', package = 'gRaphene'))
}
\keyword{map}
\keyword{raman,}
\keyword{spectrum}
