% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{gR_summary}
\alias{gR_summary}
\title{Create summary of a graphene curve fit data frame}
\usage{
gR_summary(df)
}
\arguments{
\item{df}{Computes summary (min, mean, max and sd) for \code{df}}
}
\description{
Takes the results of gRdata_read and computes summary statistics
}
\examples{
gR_summary()
}
\keyword{summary}
