% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{gR_summary_save}
\alias{gR_summary_save}
\title{Saves summary statistics for graphene curve fit}
\usage{
gR_summary_save(df_summary, filename = "summary.csv")
}
\arguments{
\item{df_summary}{Dataframe containing summary statistics}

\item{filename}{Filename of csv file (defaults to summary.csv)}
}
\description{
Saves the summary df to a csv file (default name is summary.csv)
}
\examples{
gR_summary_save()
}
\keyword{summary}
