% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRaphene-package.r
\docType{package}
\name{gRaphene}
\alias{gRaphene}
\alias{gRaphene-package}
\title{gRaphene - tools for importing and working with Raman data on graphene samples}
\description{
The data exported from Renishaw wire can be tricky to deal with. This package
collects an ensemble of functions that makes importing, modifying and plotting
these data a lot easier.
}
\details{
\href{../doc/curve_fits.html}{Curve fit vignette}

\href{../doc/spectral_maps.html}{Spectral maps vignette}
}
