% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{gRdata_map_avg}
\alias{gRdata_map_avg}
\title{Calculate average raman spectrum from a full map}
\usage{
gRdata_map_avg(df_map, normalise = FALSE)
}
\arguments{
\item{df_map}{A tidy dataframe generated by \code{gRdata_map_read()}}

\item{normalise}{Logical determining whether spectra are normalised before averaging. Defaults to FALSE.}
}
\description{
Calculates the average spectrum from a tidy map data frame
}
\examples{
gRdata_map_avg()
}
\keyword{Average,}
\keyword{map,}
\keyword{raman,}
\keyword{spectrum}
