% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{gRdata_map_read}
\alias{gRdata_map_read}
\title{Load complete spectrum map from exported txt file}
\usage{
gRdata_map_read(path)
}
\arguments{
\item{path}{Path to txt file containing the spectrum map}
}
\description{
Load all spectra and corresponding (x,y) coordinates and wrangles into a tidy tibble
}
\examples{
gRdata_map_read()
}
\keyword{map}
\keyword{raman,}
\keyword{spectrum}
