% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{gRdata_read}
\alias{gRdata_read}
\title{Load results of graphene curve fit}
\usage{
gRdata_read(path, ext = "txt",
  peaks = "^((D\\\\s)|(G\\\\s)|(2D\\\\s)|(aC\\\\s)|(DDp\\\\s)|(2Dp\\\\s)|(2Dpp\\\\s)|(SiO2\\\\s))",
  measure = "FWHM|pos|int")
}
\arguments{
\item{path}{Directory containing the data-files}

\item{ext}{Extension of data-files (default is 'txt')}

\item{peaks}{Regex string detailing the peaks to look for}

\item{measure}{Features to look for, for each peak}
}
\description{
Attempts to load all curve fits in txt files from the specified directory
}
\examples{
gRdata_read()
}
\keyword{curve}
\keyword{fit}
\keyword{raman,}
