% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{multi.fun}
\alias{multi.fun}
\title{multi.fun}
\usage{
multi.fun(x)
}
\arguments{
\item{x}{column of values for which to calculate statistics}
}
\description{
Combined function for summary statistics
}
\examples{
multi.fun(x)
}
\keyword{summary}
