% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{gr_cf_read}
\alias{gr_cf_read}
\title{Load results of graphene curve fit}
\usage{
gr_cf_read(path, ext = "txt")
}
\arguments{
\item{path}{Directory containing the data-files}

\item{ext}{Extension of data-files (default is 'txt')}
}
\description{
Attempts to load all curve fits in txt files from the specified directory
}
\details{
All files should be named as "<peak> <feature>.txt", e.g. "2D FWHM.txt" and be put in the folder specified under \code{path}.
The dataframe that is returned will contain the "<peak> <feature>" as the column names - for this to work, the filenames should only contain letters, numbers and spaces.
}
\examples{
gr_cf_read("data")
gr_cf_read(system.file('extdata/graphene_curve_fit_export', package = 'gRaphene'))
}
\seealso{
Other curve fit functions: \code{\link{gr_cf_plot_all}},
  \code{\link{gr_cf_plot}},
  \code{\link{gr_cf_summary_save}},
  \code{\link{gr_cf_summary}}
}
\keyword{curve}
\keyword{fit}
\keyword{raman,}
